requirements.txt
python-dotenv>=1.0.1
streamlit>=1.36.0
langchain>=0.2.15
langchain-openai>=0.1.22